[refactor] 기록 관련 코드 위치 변경
