+++ src/components/diary/ContentsInfo.tsx
