+++ src/components/diary/OtherInfo.tsx
