+++ src/components/diary/SayingInfo.tsx
