+++ src/components/diary/ThumbnailInfo.tsx
