+++ src/components/diary/TitleInfo.tsx
