+++ src/components/diary/WriterRateInfo.tsx
