+++ src/screens/mydiary/myExhList/MyExhListScreen.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components/native';
 import {useNavigation} from '@react-navigation/native';
 import Header from '~/components/common/Header';
 import {RootStackNavigationProp} from '~/App';
-import MyExhList from './MyExhList';
+import StoredExhList from '../../../components/diary/StoredExhList';
 import {AddMyExhButton} from '~/assets/images/index';
 
 const MyExhListScreen = () => {
@@ -22,7 +22,7 @@ const MyExhListScreen = () => {
 
       {/* body */}
       <Contents>
-        <MyExhList />
+        <StoredExhList />
       </Contents>
     </Container>
   );
